+++ react/features/shared-video/actionTypes.js
@@ -32,10 +32,10 @@ export const RESET_SHARED_VIDEO_STATUS = 'RESET_SHARED_VIDEO_STATUS';
 export const SET_DISABLE_BUTTON = 'SET_DISABLE_BUTTON';
 
 /**
- * The type of the action which signals to update owner of the shared video
+ * The type of the action which signals to send a request of broadcast of current state to the owner of the shared video
  *
  * {
- *     type: UPDATE_SHARED_VIDEO_OWNER
+ *     type: REQUEST_SHARED_VIDEO_STATE
  * }
  */
-export const UPDATE_SHARED_VIDEO_OWNER = 'UPDATE_SHARED_VIDEO_OWNER';
+export const REQUEST_SHARED_VIDEO_STATE = 'REQUEST_SHARED_VIDEO_STATE';