+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.51] - 2023-06-14
+### Changed
+- External API endpoint for requesting shared video state from video owner is added
+- External API endpoint for fetching the local participant Id is added
+
 ## [1.1.50] - 2023-06-06
 ### Changed
 - Video state update notification instead of owner update implemented